+++ CHANGELOG.md
@@ -9,7 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add the `Claude 3 Haiku` model.
+- Support `tokio::stream` for streaming API by optional.
+
+## [0.3.0] - 2024-03-14
+
+### Added
+
+- Add the Claude 3 Haiku model: `claude-3-haiku-20240307`.
 
 ## [0.2.0] - 2024-03-13
 
@@ -27,7 +33,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add the `Create a Message` API.
 
-[unreleased]: https://github.com/mochi-neko/clust/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/mochi-neko/clust/compare/v0.3.0...HEAD
+
+[0.3.0]: https://github.com/mochi-neko/clust/compare/v0.2.0...v0.3.0
 
 [0.2.0]: https://github.com/mochi-neko/clust/compare/v0.1.0...v0.2.0